Add docs to example and move to top-level
