+++ Cargo.toml
@@ -1,5 +1,6 @@
 [workspace]
 members = [
+    "examples/echo",
     "src/cli",
     "src/client",
     "src/cluster",
