+++ examples/echo/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "echo"
+version = "0.1.0"
+authors = ["Nick Mosher <nicholastmosher@gmail.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+fluvio = { path = "../../src/client" }
+async-std = "1.6.4"
